Configute DataProvider
