+++ HomeWork_16/src/test/java/MainRunnerTest.java
@@ -101,12 +101,6 @@ public class MainRunnerTest {
         // Проверяем, что возраст первого пользователя в отсортированном списке равен 60
     }
 
-    @DataProvider(name = "firstUserParams")
-    public Object[][] getFirstUserParams() {
-        return new Object[][]{
-                {"Belinda", 60} // Ожидаемое имя и возраст первого пользователя после сортировки
-        };
-    }
 
     @Test(testName = "SortUsersByNameAndAge_SeventhUser",
             dataProvider = "seventhUserParams",
@@ -129,13 +123,6 @@ public class MainRunnerTest {
                 + ", age: " + seventhUser.getAge());
     }
 
-    @DataProvider(name = "seventhUserParams")
-    public Object[][] getSeventhUserParams() {
-        return new Object[][]{
-                {"Keanu", 15} // Ожидаемое имя и возраст седьмого пользователя из списка после сортировки
-        };
-    }
-
     // Негативный тест на проверку пользователей, у которых фамилия начинается с буквы "S" или "А",
     // с возвращением пустого листа
     @Test(testName = "CheckUsersWithSorA",
@@ -152,16 +139,6 @@ public class MainRunnerTest {
             logger.error("NullPointerException occurred: " + e.getMessage());
             Assert.fail("NullPointerException occurred: " + e.getMessage());
         }
-    }
-
-    @DataProvider(name = "negativeTestData")
-    public Object[][] negativeTestData() {
-        //Возвращаем двумерный массив объектов, в котором первое значение 'null',
-        // а второй - это пустой лист, созданный при помощи коллекции 'Collections.emptyList()'
-        return new Object[][]{
-                {null},
-                {Collections.emptyList()}
-        };
     }
 
     @Test(testName = "CheckNotAllUsersOlderThan18",
